+++ src/Pkcs11Interop/LowLevelAPI8/MechanismParams/CK_PKCS5_PBKD2_PARAMS.cs
@@ -67,6 +67,6 @@ namespace Net.Pkcs11Interop.LowLevelAPI8.MechanismParams
         /// <summary>
         /// Length in bytes of the password information
         /// </summary>
-        public ulong PasswordLen;
+        public IntPtr PasswordLen;
     }
 }
